Extract editor subtitle state hook
